requirements.txt
# Overall project
flake8

# rmp_api.py
RateMyProfessorAPI 

# Database
boto3
mysql-connector-python
python-dotenv


# sentiment.py
torch == 2.4.1
transformers == 4.45.2

# frontend.py